requirements.txt
django==3.2.8
django-simple-captcha==0.5.14
markdown2==2.4.1